refactor(decorators): move to src root
