+++ src/auth/auth.controller.ts
@@ -2,7 +2,7 @@ import { BadRequestException, Body, Controller, HttpCode, InternalServerErrorExc
 import { AuthService } from './auth.service';
 import { LoginUserDto } from './dto/login-user.dto';
 import { AccessToken } from './dto/responses/access-token.dto';
-import { Public } from './decorators/public.decorator';
+import { Public } from '../decorators/public.decorator';
 import { CreateUserDto } from 'src/users/dto/create-user.dto';
 import { SanitizedUser } from 'src/users/types/sanitized-user.type';
 import { ApiBearerAuth, ApiCookieAuth, ApiResponse, ApiTags } from '@nestjs/swagger';
