+++ src/sshkeys/sshkeys.controller.ts
@@ -1,8 +1,8 @@
 import { Body, Controller, Delete, Get, HttpException, HttpStatus, Param, ParseUUIDPipe, Post } from '@nestjs/common';
 import { SshKeysService } from './sshkeys.service';
-import { AdminOnly } from 'src/auth/decorators/adminonly.decorator';
+import { AdminOnly } from 'src/decorators/adminonly.decorator';
 import { CreateSshKeyDto } from './dto/create-sshkey.dto';
-import { GetUser } from 'src/auth/decorators/user.decorator';
+import { GetUser } from 'src/decorators/user.decorator';
 import { RequestUser } from 'src/auth/types/jwt-user-data.type';
 import { ApiBearerAuth, ApiCookieAuth, ApiResponse, ApiTags } from '@nestjs/swagger';
 import { ApiStandardResponse } from 'src/interfaces/standard-response.inteface';
